% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(
  data,
  scores = c("WID_SMK_epithelial_hypoM", "WID_SMK_immune_hypoM",
    "WID_SMK_distal_epithelial_hyperM", "WID_SMK_proximal_epithelial_hyperM",
    "WID_SMK450_epithelial_hypoM", "WID_SMK450_immune_hypoM",
    "WID_SMK450_distal_epithelial_hyperM", "WID_SMK450_proximal_epithelial_hyperM")
)
}
\arguments{
\item{data}{data frame consisting of at least scores. Scores must be present in tg data frame}

\item{scores}{Scores to be normalised}
}
\value{
Data frame with additional column of scaled scores
}
\description{
Scale smoking indices, using values in TG study at M0 as reference
}
