% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WID_SMK.R
\name{WID_SMK}
\alias{WID_SMK}
\title{WID-SMK}
\usage{
WID_SMK(beta)
}
\arguments{
\item{beta}{Preprocessed data matrix}
}
\value{
Data frame with WID smk scores as columns and samples as rows
}
\description{
Key function to compute mean WID smoke scores
}
